refactor: move agent task dispatch into service
Move agent-task dispatch plan construction and durable orchestration behind a typed core service while keeping the CLI adapter focused on argument mapping and output rendering.
